refactored for routing
