+++ src/components/FindScheduleForm.jsx
@@ -1,0 +1,104 @@
+import { Redirect } from "react-router-dom";
+import React, { useState } from "react";
+import { useQuery } from "react-query";
+
+import Autocomplete from "./Autocomplete";
+import Fetch from "../common/Fetch";
+import InActiveRouteAlert from "./InActiveRouteAlert";
+import SomethingWentWrongAlert from "./SomethingWentWrongAlert";
+import { GetSuggestions } from "../common/Suggestions";
+import Suggestions from "./Suggestions";
+import { Config } from "@baltimorecounty/javascript-utilities";
+const { getValue } = Config;
+
+const FindScheduleForm = () => {
+  const [{ suggestion, status: suggestionStatus }, setSuggestion] = useState({
+    suggestion: "",
+  });
+  const [{ suggestions, status: suggestionsStatus }, setSuggestions] = useState(
+    {
+      suggestions: [],
+    }
+  );
+  const { data: addressCandidates, status } = useQuery(
+    suggestion && [
+      "getAddress",
+      {
+        endpoint: getValue("findAddressCandidates"),
+        queryString: `?address=${suggestion}`,
+      },
+    ],
+    Fetch,
+    {
+      refetchOnWindowFocus: false,
+    }
+  );
+
+  const { candidates = [] } = addressCandidates || {};
+  const hasAddressCandidates = candidates.length > 0;
+
+  const suggest = async (query, populateResults) => {
+    setSuggestions({});
+    try {
+      const suggestions = await GetSuggestions(query);
+      populateResults(suggestions);
+    } catch (ex) {
+      setSuggestion({
+        suggestion: null,
+        status: "error",
+      });
+    }
+  };
+
+  const handleValueSelect = (selectedValue) => {
+    setSuggestion({
+      suggestion: selectedValue,
+      status: "success",
+    });
+    setSuggestions({});
+  };
+
+  const handleSubmit = async (submitEvent) => {
+    submitEvent.preventDefault();
+    const addressQuery = document.getElementById("address-lookup").value;
+
+    const suggestions = await GetSuggestions(addressQuery);
+
+    setSuggestions({
+      suggestions,
+      status: suggestions && suggestions.length > 0 ? "success" : "error",
+    });
+  };
+
+  if ([status, suggestionStatus].some((x) => x === "error")) {
+    return <SomethingWentWrongAlert />;
+  }
+
+  if (suggestionsStatus === "error") {
+    return <InActiveRouteAlert />;
+  }
+
+  if (hasAddressCandidates) {
+    return <Redirect to={`/${candidates[0].attributes.placeName}`} />;
+  }
+
+  return (
+    <div className="App">
+      <form onSubmit={handleSubmit}>
+        <Autocomplete
+          id="address-lookup"
+          label="Find Your Collection Schedule"
+          suggest={suggest}
+          onConfirm={handleValueSelect}
+          minLength={3}
+        />
+        <button type="submit">Submit</button>
+      </form>
+      {suggestions && suggestions.length > 0 && (
+        <Suggestions suggestions={suggestions} />
+      )}
+    </div>
+  );
+};
+
+export default FindScheduleForm;
